This is really two changes in one but given interdependencies and expected review latency I needed to combine the CLs.
1. Made the 5* toolpath build using the go build
   system. Hooked the subdirectories to clean.bash but added a
   separate make5.bash for now. Minor massage to make the code
   more similar to the current structure of 6c/6a/6l.

2. Change all references from long to int32 in line with
   similar change for the other toolchains.

The end result is that 5c, 5a and 5l can now be compiled and
the executables start up properly. Haven't thrown any input at
them yet.

R=rsc
APPROVED=rsc
DELTA=1052  (392 added, 328 deleted, 332 changed)
OCL=26757
CL=26761

+++ src/cmd/5c/cgen.c
@@ -37,7 +37,7 @@ cgen(Node *n, Node *nn, int inrel)
 	Prog *p1;
 	Node nod, nod1, nod2, nod3, nod4;
 	int o, t;
-	long v, curs;
+	int32 v, curs;
 
 	if(debug['g']) {
 		prtree(nn, "cgen lhs");
@@ -475,7 +475,7 @@ cgen(Node *n, Node *nn, int inrel)
 				diag(n, "DOT and no offset");
 				break;
 			}
-			nod.xoffset += (long)r->vconst;
+			nod.xoffset += (int32)r->vconst;
 			nod.type = n->type;
 			cgen(&nod, nn, 0);
 		}
@@ -590,7 +590,7 @@ void
 reglcgen(Node *t, Node *n, Node *nn)
 {
 	Node *r;
-	long v;
+	int32 v;
 
 	regialloc(t, n, nn);
 	if(n->op == OIND) {
@@ -705,7 +705,7 @@ boolgen(Node *n, int true, Node *nn)
 	int o;
 	Prog *p1, *p2;
 	Node *l, *r, nod, nod1;
-	long curs;
+	int32 curs;
 
 	if(debug['g']) {
 		prtree(nn, "boolgen lhs");
@@ -864,12 +864,12 @@ boolgen(Node *n, int true, Node *nn)
 }
 
 void
-sugen(Node *n, Node *nn, long w)
+sugen(Node *n, Node *nn, int32 w)
 {
 	Prog *p1;
 	Node nod0, nod1, nod2, nod3, nod4, *l, *r;
 	Type *t;
-	long pc1;
+	int32 pc1;
 	int i, m, c;
 
 	if(n == Z || n->type == T)
@@ -929,7 +929,7 @@ sugen(Node *n, Node *nn, long w)
 				diag(n, "DOT and no offset");
 				break;
 			}
-			nod1.xoffset += (long)r->vconst;
+			nod1.xoffset += (int32)r->vconst;
 			nod1.type = n->type;
 			sugen(&nod1, nn, w);
 		}
