+++ src/cmd/5l/list.c
@@ -156,7 +156,7 @@ Dconv(Fmt *fp)
 	char str[STRINGSZ];
 	char *op;
 	Adr *a;
-	long v;
+	int32 v;
 
 	a = va_arg(fp->args, Adr*);
 	switch(a->type) {
@@ -335,7 +335,7 @@ Sconv(Fmt *fp)
 
 	a = va_arg(fp->args, char*);
 	p = str;
-	for(i=0; i<sizeof(long); i++) {
+	for(i=0; i<sizeof(int32); i++) {
 		c = a[i] & 0xff;
 		if(c >= 'a' && c <= 'z' ||
 		   c >= 'A' && c <= 'Z' ||
